fix: pickBiasedSp (int) 캐스팅 오버플로우 수정
span * biased 결과를 (int)로 캐스팅해 2.1억 초과 시 음수 오버플로우 발생.
Math.round(double)은 이미 long 반환 — (int) 제거로 해결.

+++ src/main/java/my/prac/api/loa/controller/BossAttackController.java
@@ -8374,11 +8374,11 @@ public class BossAttackController {
 
 	
 	private SP pickBiasedSp(long min, long max) {
-	    double r = ThreadLocalRandom.current().nextDouble(); 
-	    double biased = Math.pow(r, 6); 
+	    double r = ThreadLocalRandom.current().nextDouble();
+	    double biased = Math.pow(r, 6);
 
 	    long span = max - min;
-	    long value = min + (int)Math.round(span * biased);
+	    long value = min + Math.round(span * biased); // (int) 캐스팅 제거 — int 범위 초과 시 오버플로우 방지
 	    SP sp = SP.fromSp(value);
 	    return sp;
 	}
